Поправил инициализацию ящиков
